2008-02-17 Erik Hjortsberg  <erik.hjortsberg@iteam.se>
	* pretty much every source file in the project: include config.h if available and fix up licensing and headers

+++ src/components/ogre/input/Input.h
@@ -326,7 +326,7 @@ protected:
 	the amount of time since the last right mouse click
 	used for detecting double clicks
 	*/
-	uint32_t mTimeSinceLastRightMouseClick;
+	float mTimeSinceLastRightMouseClick;
 	
 	uint32_t mLastTick;
